fix: model waitingCount as integer in wait state statistics schema
Use z.number().int() for waitingCount to match the count-modeling convention
in this package (e.g. operationsTotalCount in batch-operation.ts). Plain
z.number() would also accept floats, weakening runtime validation and the
inferred type. Addresses PR review feedback.

+++ webapp/client/packages/camunda-api-zod-schemas/lib/8.10/process-instance.ts
@@ -235,7 +235,7 @@ const getProcessInstanceWaitStateStatistics: Endpoint<Pick<ProcessInstance, 'pro
 
 const waitStateStatisticSchema = z.object({
 	elementId: z.string(),
-	waitingCount: z.number(),
+	waitingCount: z.number().int(),
 });
 type WaitStateStatistic = z.infer<typeof waitStateStatisticSchema>;
